rename bittensor/types.py to bittensor/bt_types.py to prevent mixups after pip install
Although it seems common to have types.py (e.g. scalecodec, torch,
substrateinterface) this may lead to issues after applying pip install
-e to a package, as is suggested for bittensor (see git grep 'pip
install -e.* bittensor')

Issues were observed where circular imports would arise as Python's
native typing.py would include bittensor's types.py.
